+++ x/params/simulation/operations_test.go
@@ -37,7 +37,7 @@ func (pc MockParamChange) SimValue() simtypes.SimValFn {
 }
 
 // make sure that the MockParamChange satisfied the ParamChange interface
-var _ simtypes.ParamChange = MockParamChange{}
+var _ simtypes.LegacyParamChange = MockParamChange{}
 
 func TestSimulateParamChangeProposalContent(t *testing.T) {
 	s := rand.NewSource(1)
@@ -45,7 +45,7 @@ func TestSimulateParamChangeProposalContent(t *testing.T) {
 
 	ctx := sdk.NewContext(nil, tmproto.Header{}, true, nil)
 	accounts := simtypes.RandomAccounts(r, 3)
-	paramChangePool := []simtypes.ParamChange{MockParamChange{1}, MockParamChange{2}, MockParamChange{3}}
+	paramChangePool := []simtypes.LegacyParamChange{MockParamChange{1}, MockParamChange{2}, MockParamChange{3}}
 
 	// execute operation
 	op := simulation.SimulateParamChangeProposalContent(paramChangePool)